move batch filter to fluent
