+++ great_expectations/datasource/fluent/data_asset/path/directory_asset.py
@@ -33,8 +33,8 @@ from great_expectations.datasource.fluent.interfaces import DatasourceT, Partiti
 if TYPE_CHECKING:
     from great_expectations.alias_types import PathStr
     from great_expectations.core.batch_definition import BatchDefinition
-    from great_expectations.datasource.data_connector.batch_filter import BatchSlice
     from great_expectations.datasource.fluent import BatchParameters
+    from great_expectations.datasource.fluent.data_connector.batch_filter import BatchSlice
 
 
 class DirectoryDataAsset(PathDataAsset[DatasourceT, ColumnPartitioner], Generic[DatasourceT], ABC):
